Move reqif utils to a dedicated class - (Task #807)
